Make a test_utils.rs module and move test_assembler::Section extensions to it from cfi.rs's `tests` module
