+++ src/cfi.rs
@@ -1806,7 +1806,6 @@ impl<'input, Endian> FallibleIterator for CallFrameInstructionIter<'input, Endia
 
 #[cfg(test)]
 mod tests {
-    extern crate leb128;
     extern crate test_assembler;
 
     use super::*;
@@ -1815,14 +1814,11 @@ mod tests {
     use endianity::{BigEndian, Endianity, EndianBuf, LittleEndian, NativeEndian};
     use parser::{Error, Format, Result};
     use self::test_assembler::{Endian, Label, LabelMaker, LabelOrNum, Section, ToLabelOrNum};
+    use test_util::GimliSectionMethods;
 
     // Mixin methods for `Section` to help define binary test data.
 
-    trait CfiSectionMethods {
-        fn e32<'a, T>(self, endian: Endian, val: T) -> Self where T: ToLabelOrNum<'a, u32>;
-        fn e64<'a, T>(self, endian: Endian, val: T) -> Self where T: ToLabelOrNum<'a, u64>;
-        fn sleb(self, val: i64) -> Self;
-        fn uleb(self, val: u64) -> Self;
+    trait CfiSectionMethods: GimliSectionMethods {
         fn cie<'input, E>(self,
                           endian: Endian,
                           cie: &mut CommonInformationEntry<'input, E>)
@@ -1838,36 +1834,6 @@ mod tests {
     }
 
     impl CfiSectionMethods for Section {
-        fn e32<'a, T>(self, endian: Endian, val: T) -> Self
-            where T: ToLabelOrNum<'a, u32>
-        {
-            match endian {
-                Endian::Little => self.L32(val),
-                Endian::Big => self.B32(val),
-            }
-        }
-
-        fn e64<'a, T>(self, endian: Endian, val: T) -> Self
-            where T: ToLabelOrNum<'a, u64>
-        {
-            match endian {
-                Endian::Little => self.L64(val),
-                Endian::Big => self.B64(val),
-            }
-        }
-
-        fn sleb(self, val: i64) -> Self {
-            let mut buf = Vec::new();
-            let written = leb128::write::signed(&mut buf, val).unwrap();
-            self.append_bytes(&buf[0..written])
-        }
-
-        fn uleb(self, val: u64) -> Self {
-            let mut buf = Vec::new();
-            let written = leb128::write::unsigned(&mut buf, val).unwrap();
-            self.append_bytes(&buf[0..written])
-        }
-
         fn cie<'input, E>(self, endian: Endian, cie: &mut CommonInformationEntry<'input, E>) -> Self
             where E: Endianity
         {
